backend/requirements.txt
# One Logikality backend dependencies
# Version pins match Title Intelligence Hub where the same package is used;
# see docs/TechStack.md §3–9 for rationale. AI / PDF / fuzzy-match / Playwright
# dependencies land in Phase 3 (Step 4) when the ECV pipeline is wired.

# Core framework
fastapi==0.115.14
uvicorn[standard]==0.34.3
python-multipart==0.0.22
python-dotenv==1.2.2
pydantic[email]==2.12.5
pydantic-settings>=2.0

# Database
# The [asyncio] extra pulls in greenlet, which SQLAlchemy's async engine
# requires. On Apple Silicon the plain SQLAlchemy wheel does not install
# greenlet transitively, so request the extra explicitly.
SQLAlchemy[asyncio]==2.0.48
asyncpg==0.30.0
alembic==1.14.1

# Auth & Security (password hashing lands in Step 1; present now so auth is one pip install away)
PyJWT==2.11.0
bcrypt==4.0.1
passlib[bcrypt]>=1.7

# HTTP & file I/O
# httpx 0.28 pulled in transitively by anthropic>=0.40; pinning to match.
httpx==0.28.1
aiofiles==24.1.0

# PDF export (US-8.1). Pure-Python; no system deps, no wkhtmltopdf / Chrome
# shim to maintain. Good enough for a stakeholder-facing validation report
# with tables + brand-coloured chrome.
reportlab==4.2.5

# Workflow orchestration
temporalio>=1.7

# AI providers (Phase 3 / Step 4 — ECV pipeline)
# google-genai is Google's unified SDK for Gemini via both the Developer API
# and Vertex AI; we only use the vertexai=True code path per CLAUDE.md.
anthropic>=0.40
google-genai>=0.3

# PDF ingestion (page splitting + text extraction for the classify stage).
# Form-OCR for scanned PDFs lands in M3 via Document AI.
pypdf>=5.0

# Test dependencies
pytest>=9.0
pytest-asyncio>=0.21
aiosqlite>=0.20

# Linting / formatting (upgrade over TI Hub per docs/TechStack.md §10)
ruff>=0.6
